Move project_info.py to bot package to avoid defining names on the top level that could clash with any client package with the same name. Libraries should have all their modules namespaced.
